Add LaRecipe library to handle documentation
The LaRecipe library has been added to improve the way the application manages documentation. This library allows generating documents using markdown, which simplifies the editing process while maintaining a cleaner and clearer structure of the files. It also allows configuring different settings for documentation such as versions and route paths, which adds an extra layer of control and flexibility.

Several changes have been made to other modules to keep the code clean and structured. These include changes in FamilyController and User model. The function signatures of FamilyController have been updated to include explicit return types. Import of unnecessary relationship in the Family model has been removed.

Lastly, there are some slight alterations to the formatting in RoleResource and AdminPanelProvider to comply with the pre-set coding standard.